Refactored logout to use the onSuccess and onError inside the API hook instead of in a new hook.
